+++ sdk/src/vite/manifestPlugin.mts
@@ -1,8 +1,28 @@
 import { readFile } from "node:fs/promises";
-import { type Plugin } from "vite";
+import { type Plugin, type ViteDevServer } from "vite";
 
 const virtualModuleId = "virtual:manifest.js";
 const resolvedVirtualModuleId = "\0" + virtualModuleId;
+
+const getCssForModule = (
+  server: ViteDevServer,
+  moduleId: string,
+  css: Set<string>,
+) => {
+  const moduleNode = server.moduleGraph.getModuleById(moduleId);
+
+  if (!moduleNode) {
+    return;
+  }
+
+  for (const importedModule of moduleNode.importedModules) {
+    if (importedModule.url.endsWith(".css")) {
+      css.add(importedModule.url);
+    }
+
+    getCssForModule(server, importedModule.id!, css);
+  }
+};
 
 export const manifestPlugin = ({
   manifestPath,
@@ -30,6 +50,38 @@ export const manifestPlugin = ({
         const manifestContent = await readFile(manifestPath, "utf-8");
         return `export default ${manifestContent}`;
       }
+    },
+    configureServer(server) {
+      server.middlewares.use("/__rwsdk_manifest", async (req, res, next) => {
+        try {
+          const manifest: Record<string, { file: string; css: string[] }> = {};
+
+          for (const file of server.moduleGraph.fileToModulesMap.keys()) {
+            const modules = server.moduleGraph.getModulesByFile(file);
+
+            if (!modules) {
+              continue;
+            }
+
+            for (const module of modules) {
+              if (module.file) {
+                const css = new Set<string>();
+                getCssForModule(server, module.id!, css);
+
+                manifest[module.file.replace(server.config.root, "")] = {
+                  file: module.url,
+                  css: Array.from(css),
+                };
+              }
+            }
+          }
+
+          res.setHeader("Content-Type", "application/json");
+          res.end(JSON.stringify(manifest));
+        } catch (e) {
+          next(e);
+        }
+      });
     },
   };
 };
