+++ sdk/src/vite/redwoodPlugin.mts
@@ -27,7 +27,6 @@ import { ssrBridgePlugin } from "./ssrBridgePlugin.mjs";
 import { hasPkgScript } from "../lib/hasPkgScript.mjs";
 import { devServerTimingPlugin } from "./devServerTimingPlugin.mjs";
 import { manifestPlugin } from "./manifestPlugin.mjs";
-import { devManifestPlugin } from "./devManifestPlugin.mjs";
 
 export type RedwoodPluginOptions = {
   silent?: boolean;
@@ -162,7 +161,6 @@ export const redwoodPlugin = async (
         "manifest.json",
       ),
     }),
-    devManifestPlugin(),
     moveStaticAssetsPlugin({ rootDir: projectRootDir }),
     prismaPlugin({ projectRootDir }),
   ];
